+++ src/ocean/util/container/queue/FlexibleFileQueue.d
@@ -219,8 +219,7 @@ public class FlexibleFileQueue : IByteQueue
         this.size = size;
         this.open_existing = open_existing;
 
-        if ( this.open_existing && Filesystem.exists(this.path) &&
-            Filesystem.exists(this.index_path) )
+        if ( this.open_existing && this.exists(this.path) )
         {
             this.file_out = new File(this.path, File.WriteAppending);
             this.file_index = new File(this.index_path, File.ReadWriteExisting);
@@ -243,6 +242,30 @@ public class FlexibleFileQueue : IByteQueue
         this.ext_in.seek(this.file_out.length - this.bytes_in_file);
 
         this.files_open = true;
+    }
+
+
+    /***************************************************************************
+
+        Checks whether a file and the associated index file exist at the
+        specified path.
+
+        Note that this function allocates on every call! It is only intended to
+        be called during application startup.
+
+        Params:
+            path = path to check (".index" is appended to check for the
+                corresponding index file)
+
+        Returns:
+            true if the specified file and index file exist
+
+    ***************************************************************************/
+
+    public static bool exists ( cstring path )
+    {
+        return Filesystem.exists(path)
+            && Filesystem.exists(path ~ IndexExtension);
     }